+++ src/cpu/src/op/clear.c
@@ -20,7 +20,7 @@ extern "C" {
 #include "mod.h"
 #include "params.h"
 
-extern volatile bool_t _read_fpga_info;
+extern volatile bool_t _read_fpga_state;
 extern volatile uint16_t _fpga_flags_internal;
 
 extern volatile uint32_t _mod_cycle;
@@ -45,7 +45,7 @@ uint8_t clear(void) {
   _mod_freq_div = 5120;
   _stm_freq_div = 0xFFFFFFFF;
 
-  _read_fpga_info = false;
+  _read_fpga_state = false;
 
   _fpga_flags_internal = 0;
   bram_write(BRAM_SELECT_CONTROLLER, BRAM_ADDR_CTL_FLAG, _fpga_flags_internal);