+++ src/cpu/tests/op/reads_fpga_state.cpp
@@ -1,4 +1,4 @@
-// File: reads_fpga_info.cpp
+// File: reads_fpga_state.cpp
 // Project: op
 // Created Date: 31/12/2023
 // Author: Shun Suzuki
@@ -20,7 +20,7 @@ extern "C" {
 extern TX_STR _sTx;
 }
 
-TEST(Op, ReadsFPGAInfo) {
+TEST(Op, ReadsFPGAState) {
   init_app();
 
   RX_STR data;
@@ -32,7 +32,7 @@ TEST(Op, ReadsFPGAInfo) {
     header->slot_2_offset = 0;
 
     auto* data_body = reinterpret_cast<uint8_t*>(data.data) + sizeof(Header);
-    data_body[0] = TAG_READS_FPGA_INFO;
+    data_body[0] = TAG_READS_FPGA_STATE;
     data_body[1] = 0x01;  // enable
 
     auto frame = to_frame_data(data);
@@ -45,7 +45,7 @@ TEST(Op, ReadsFPGAInfo) {
     ASSERT_EQ(0x00, _sTx.ack & 0xFF);
 
     update();
-    ASSERT_EQ(READS_FPGA_INFO_ENABLED, _sTx.ack & 0xFF);
+    ASSERT_EQ(READS_FPGA_STATE_ENABLED, _sTx.ack & 0xFF);
   }
 
   {
@@ -53,7 +53,7 @@ TEST(Op, ReadsFPGAInfo) {
     header->slot_2_offset = 0;
 
     auto* data_body = reinterpret_cast<uint8_t*>(data.data) + sizeof(Header);
-    data_body[0] = TAG_READS_FPGA_INFO;
+    data_body[0] = TAG_READS_FPGA_STATE;
     data_body[1] = 0x00;  // disable
 
     auto frame = to_frame_data(data);
@@ -63,7 +63,7 @@ TEST(Op, ReadsFPGAInfo) {
 
     const auto ack = _sTx.ack >> 8;
     ASSERT_EQ(ack, header->msg_id);
-    ASSERT_EQ(READS_FPGA_INFO_ENABLED, _sTx.ack & 0xFF);
+    ASSERT_EQ(READS_FPGA_STATE_ENABLED, _sTx.ack & 0xFF);
 
     update();
     ASSERT_EQ(0x00, _sTx.ack & 0xFF);